Per-user data isolation, profile form improvements, Supabase sync
User isolation:
- Clear all local data (closet, outfits, week plan, session) on SIGNED_OUT
- Reset AppSessionProvider state on sign-out; re-hydrate from backend on new sign-in
- New accounts now start with a clean slate and land on onboarding

Supabase sync infrastructure:
- lib/supabase-data.ts: CRUD helpers for closet_items, saved_outfits, week_plan
- lib/user-data-sync.ts: syncUserDataOnSignIn (cloud-wins), clearAllLocalUserData
- closet/outfits/week storage: fire-and-forget upsert/delete to Supabase on writes

Profile form:
- Height supports cm / ft+in with live conversion
- Weight supports kg / lbs with live conversion
- heightCm/weightKg changed to Float in Prisma schema + migration
- Removed 'prefer-not-to-say' gender option (merged into non-binary)
- Added 'black' skin tone option
- SelectorGrid / SelectorCard components for image-based option pickers

Other:
- Fix app/index.tsx routing (was hardcoded to onboarding for testing)
- GoogleService-Info.plist added for Firebase initialization
- Style tier images added for menswear + womenswear

+++ backend/src/modules/profile/profile.validation.ts
@@ -2,8 +2,8 @@ import { z } from 'zod';
 
 export const upsertProfileSchema = z.object({
   gender: z.string().min(1),
-  heightCm: z.number().int().positive(),
-  weightKg: z.number().int().positive(),
+  heightCm: z.number().positive(),
+  weightKg: z.number().positive(),
   fitPreference: z.string().min(1),
   stylePreference: z.string().min(1),
   budget: z.string().min(1),
